feat(dart): support nested container field codec for dart (#3641)
## Why?

Dart needs a canonical field type annotation surface that can describe
nested container element/key/value semantics, including fixed-width and
encoded numeric types, without relying on separate numeric wrapper
annotations. The schema fingerprint calculation also needs to include
nested container type IDs consistently across runtimes so
compatible/schema-version checks agree for annotated nested containers.

## What does this PR do?

- Adds Dart nested container field type support through `ForyField(type:
...)`, `MapField`, `ListType`, scalar type specs, and generated support
for nested element/key/value metadata.
- Updates the Dart compiler generator and generated test models to emit
Dart `int` fields with explicit Fory type specs for fixed-width integer
semantics, while keeping exact-width wrapper classes where the Dart
runtime still requires them.
- Removes obsolete Dart numeric wrapper exports/tests for small
fixed-width integer wrappers and adds runtime validation utilities for
integer bounds.
- Extends Dart cross-language and unit coverage for nested annotated
containers, integer round trips, type-spec annotations, signed/unsigned
serializers, and xlang protocol behavior.
- Aligns nested container schema fingerprint generation across Java,
C++, C#, Go, JavaScript, Python, Rust, and Swift, and adds focused
Swift/Java xlang coverage for Dart nested annotated containers.
- Updates Dart guides, package README content, generated-code docs, and
the xlang serialization spec to document the new annotation/type mapping
model.

## Related issues



## AI Contribution Checklist



- [ ] Substantial AI assistance was used in this PR: `yes` / `no`
- [ ] If `yes`, I included a completed [AI Contribution
Checklist](https://github.com/apache/fory/blob/main/AI_POLICY.md#9-contributor-checklist-for-ai-assisted-prs)
in this PR description and the required `AI Usage Disclosure`.
- [ ] If `yes`, my PR description includes the required `ai_review`
summary and screenshot evidence of the final clean AI review results
from both fresh reviewers on the current PR diff or current HEAD after
the latest code changes.



## Does this PR introduce any user-facing change?



- [x] Does this PR introduce any public API change?
- [ ] Does this PR introduce any binary protocol compatibility change?

This changes the Dart annotation/type-spec surface for fixed-width
numeric and nested container field declarations. It updates schema
fingerprint calculation to include nested container metadata
consistently, but does not intentionally change the xlang binary
protocol format.

## Benchmark

+++ dart/packages/fory-test/test/compatible_struct_slots_test.dart
@@ -44,13 +44,13 @@ void main() {
     );
 
     final original = SimpleStruct()
-      ..f2 = Int32(1)
-      ..f7 = Int32(2)
-      ..f8 = Int32(3)
-      ..last = Int32(4)
+      ..f2 = (1)
+      ..f7 = (2)
+      ..f8 = (3)
+      ..last = (4)
       ..f4 = 'outer'
       ..f6 = <String>['a', 'b']
-      ..f1 = <Int32?, double?>{Int32(7): 9.5}
+      ..f1 = <int?, double?>{(7): 9.5}
       ..f3 = (Item()..name = 'inner')
       ..f5 = Color.blue;
 
@@ -62,13 +62,13 @@ void main() {
     final secondBytes = fory.serialize(decoded);
     final roundTrip = fory.deserialize<SimpleStruct>(secondBytes);
 
-    expect(roundTrip.f2, equals(Int32(1)));
-    expect(roundTrip.f7, equals(Int32(2)));
-    expect(roundTrip.f8, equals(Int32(3)));
-    expect(roundTrip.last, equals(Int32(4)));
+    expect(roundTrip.f2, equals((1)));
+    expect(roundTrip.f7, equals((2)));
+    expect(roundTrip.f8, equals((3)));
+    expect(roundTrip.last, equals((4)));
     expect(roundTrip.f4, equals('outer'));
     expect(roundTrip.f6, equals(<String>['a', 'b']));
-    expect(roundTrip.f1[Int32(7)], equals(9.5));
+    expect(roundTrip.f1[(7)], equals(9.5));
     expect(roundTrip.f3.name, equals('inner'));
     expect(roundTrip.f5, equals(Color.blue));
   });
